requirements.txt
opencv-contrib-python
pyzbar
pandas
openpyxl
Pillow
